+++ app/src/main/java/com/martin/preventapp/view/entities/ProductPriceAudit.kt
@@ -3,7 +3,7 @@ package com.martin.preventapp.view.entities
 data class ProductPriceAudit(
     val productName: String,
     val listName: String,
-    val unitPrice: Int,
+    val unitPrice: Float,
     val validityDate: String,
     val creationDate: String
 )
